+++ src/module/linksHere.js
@@ -6,7 +6,7 @@
 import { _analytics } from './_analytics'
 import { $progress, $link } from './_elements'
 import { _msg } from './_msg'
-import { mwApi, mwConfig } from './mw'
+import { mwApi, mwConfig } from '../utils/mw'
 import { quickEdit } from './quickEdit'
 
 /**
