--- src/module/test.js
@@ -1,3 +1,0 @@
-export function test() {
-  console.log('InPageEdit is runing version', window.InPageEdit.version)
-}
